requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
langchain-google-genai==1.0.6
langgraph==0.1.4
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.4
numpy==1.26.4
